+++ demo/index.html
@@ -5,7 +5,7 @@
     <title>SQLite Wasm Demo</title>
     <script type="module" src="script.js"></script>
     <script type="module" src="main-thread.js"></script>
-    <script type="module" src="sqlite.client.js"></script>
+    <script type="module" src="sqlite-client.js"></script>
   </head>
   <body>
     <h1>SQLite Wasm Demo</h1>
