+++ demo/sqlite-client.js
@@ -10,7 +10,9 @@ for (let i = 20; i <= 25; ++i) {
   await sqliteClient.executeSql('INSERT INTO t(a,b) VALUES (?,?)', [i, i * 2]);
 }
 
-const rows = await sqliteClient.executeSql('SELECT a FROM t ORDER BY a LIMIT 3');
+const rows = await sqliteClient.executeSql(
+  'SELECT a FROM t ORDER BY a LIMIT 3',
+);
 
 console.log(rows);
